[UPLOADS]_simple_example_of_parser_and_uploader_to_BQ/requirements.txt
Flask==2.0.3
telebot==0.0.4
pyTelegramBotAPI==3.8.3
gunicorn==19.9.0
Werkzeug==2.2.2
python-dotenv==1.0.1
pandas==1.0.3
pandas-gbq==0.17.4
google-cloud-bigquery==1.28.0
numpy==1.23.1
protobuf==3.20.1
google-api-core==1.33.2
requests==2.31.0
selenium==4.10.0
beautifulsoup4==4.9.3
webdriver-manager==4.0.1
chromedriver-autoinstaller==0.6.4


